separate declassing vs unittests
